chg:usr: Rename Epic flags from -E to --epic
= This is a breaking change!

This also changes the epic formatting designation to `%epic`, iteration
flag to `-i`, iteration formatting designation to `%i`, and story ID
formatting designation to `%id`.

Since there was a class between estimate and epic for the -e flag and
formatting designation, I've chosen to use the simpler flag/formatter
for the field closer to the story - the estimate.

This also reserves capital letter flags for side-effect or abnormal
actions (opening the browser, downloading files, etc).