add:usr:config get mentionName and urlSlug from API
Also adds caching of the config instead of reading the file
everytime something calls `loadConfig()`.

I moved the config checks in loadConfig to have things more
centralized. The loadCachedConfig function can be used in the
rare cases where we need to read the config without triggering
the checks (e.g. the first time we fetch the token)