fix #293 retrieval of project and epic entities
When user is creating a story and pass epic and project ids (instead of name),
shortcut-cli was not able to associate the project (or epic) with the given id.
The reason is that projects and epics are identified by numbers when user is giving
as input a string.
This fix implements a naive solution that is to locate for the project/epic by string
and then by number after a cast.